Move validation code to a separate file
